+++ python/sglang/srt/managers/rust_server.py
@@ -18,7 +18,7 @@ from typing import TYPE_CHECKING, Any, List, Optional, Tuple
 
 import msgspec
 
-from sglang.srt.managers.io_struct import (
+from sglang.srt.managers.utils import (
     MsgpackDecodeError,
     msgpack_decode_explained,
 )
